move the parseFilter bits into aws/adapter with the other filter parsing bits
