+++ aws/adapter_test.go
@@ -767,3 +767,76 @@ func TestFindLBSubnets(tt *testing.T) {
 		})
 	}
 }
+
+func TestParseFilterTagsDefault(t *testing.T) {
+        for _, test := range []struct {
+                name          string
+                clusterId     string
+                want          map[string][]string
+        }{
+                {
+                        "success-default-filter-asg",
+                        "mycluster",
+                        map[string][]string{
+                                clusterIDTagPrefix + "mycluster": []string{resourceLifecycleOwned},
+                        },
+                },
+        } {
+                t.Run(fmt.Sprintf("%v", test.name), func(t *testing.T) {
+                        got := parseAutoscaleFilterTags(test.clusterId)
+
+                        if !reflect.DeepEqual(test.want, got) {
+                                t.Errorf("unexpected result. wanted %+v, got %+v", test.want, got)
+                        }
+                })
+        }
+}
+
+func TestParseFilterTagsCustom(t *testing.T) {
+        for _, test := range []struct {
+                name          string
+                clusterId     string
+                customFilter  string
+                want          map[string][]string
+        }{
+                {
+                        "success-custom-filter-asg",
+                        "mycluster",
+                        "tag:kubernetes.io/cluster/mycluster=owned tag-key=k8s.io/role/node tag-key=custom.com/ingress",
+                        map[string][]string{
+                                "kubernetes.io/cluster/mycluster": []string{"owned"},
+                                "k8s.io/role/node": []string{},
+                                "custom.com/ingress": []string{},
+                        },
+                },
+                {
+                        "success-custom-filter-multivalue-asg",
+                        "mycluster",
+                        "tag:kubernetes.io/cluster/mycluster=owned tag-key=k8s.io/role/node tag:custom.com/ingress=owned,shared",
+                        map[string][]string{
+                                "kubernetes.io/cluster/mycluster": []string{"owned"},
+                                "k8s.io/role/node": []string{},
+                                "custom.com/ingress": []string{"owned", "shared"},
+                        },
+                },
+                {
+                        "success-custom-filter-fallback-to-default-asg",
+                        "mycluster",
+                        "tag:goodtag=foo tag-key=alsogood thisisabadtag andthisonetoo",
+                        map[string][]string{
+                                clusterIDTagPrefix + "mycluster": []string{resourceLifecycleOwned},
+                        },
+                },
+        } {
+                t.Run(fmt.Sprintf("%v", test.name), func(t *testing.T) {
+                        os.Setenv("CUSTOM_FILTERS", test.customFilter)
+                        got := parseAutoscaleFilterTags(test.clusterId)
+                        os.Unsetenv("CUSTOM_FILTERS")
+
+                        if !reflect.DeepEqual(test.want, got) {
+                                t.Errorf("unexpected result. wanted %+v, got %+v", test.want, got)
+                        }
+                })
+        }
+}
+
